Code Smell - refactoring GroupService
